// src/app/core/shared/models/product.model.ts
export interface Product{
    product_id: number | any;
    vendor_id: number | any;
    business_id: number | any;
    business_name: string | any;
    product_name: string | any;
    description: string | any;
    price: number | any;
    brand: string | any;
    quantity: number | any;
    category: string | any;
    path: string | any;
    registration_date: string | any;
}